+++ scripts/setup-ts-cache.js
@@ -7,9 +7,7 @@ const zlib = require(`zlib`);
 
 // Needed by the worker spawned by esbuild
 if (process.versions.pnp)
-  // Unquoted because Yarn doesn't support it quoted yet
-  // TODO: make Yarn support quoted PnP requires in NODE_OPTIONS
-  process.env.NODE_OPTIONS = `${process.env.NODE_OPTIONS || ``} --require ${require.resolve(`pnpapi`)}`;
+  process.env.NODE_OPTIONS = `${process.env.NODE_OPTIONS || ``} --require ${JSON.stringify(require.resolve(`pnpapi`))}`;
 
 const resolveVirtual = process.versions.pnp
   ? require(`pnpapi`).resolveVirtual